zed/requirements.txt
numpy
scipy
pandas
matplotlib
pyadi-iio
pylibiio